LR_4/requirements.txt
asgiref==3.8.1
backcall==0.2.0
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
docopt==0.6.2
drf-api-logger==1.1.21
fastjsonschema==2.21.2
idna==3.10
oauthlib==3.2.2
pickleshare==0.7.5
pillow==11.3.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pyzmq==27.1.0
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.8
sqlparse==0.5.3
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
